requirements.txt
tensorflow
keras
numpy
pandas
scikit-learn
matplotlib
seaborn
Pillow
flask
gunicorn